requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
astroid==3.3.9
    # via pylint
bcrypt==4.2.1
    # via -r requirements.in
beautifulsoup4==4.13.3
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
coverage==7.6.12
    # via -r requirements.in
dill==0.3.9
    # via pylint
fastapi==0.115.12
    # via -r requirements.in
filelock==3.18.0
    # via tldextract
greenlet==3.1.1
    # via playwright
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   tldextract
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via pylint
jinja2==3.1.6
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via pylint
packaging==24.2
    # via pytest
platformdirs==4.3.7
    # via pylint
playwright==1.51.0
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.8.1
    # via -r requirements.in
pyee==12.1.1
    # via playwright
pylint==3.3.6
    # via -r requirements.in
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.26.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via
    #   requests-file
    #   tldextract
requests-file==2.1.0
    # via tldextract
ruff==0.9.10
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
starlette==0.46.1
    # via fastapi
tldextract==5.1.3
    # via -r requirements.in
tomlkit==0.13.2
    # via pylint
typing-extensions==4.12.2
    # via
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pyee
urllib3==2.3.0
    # via requests
uvicorn[standard]==0.34.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
validators==0.34.0
    # via -r requirements.in
watchfiles==1.0.4
    # via uvicorn
websockets==15.0.1
    # via uvicorn
